segmentation/mask2former/requirements.txt
torch==2.0.1
torchvision==0.15.2
timm==0.5.4
mmcv==2.0.0
mmengine==0.7.3
mmsegmentation==1.0.0
mmdet==3.0.0
numpy==1.26.4
matplotlib>=3.5.0
yapf==0.32.0
